don't reference root in the startup screen
It seems that I can pass in -u user and that will override the existing
-u root.
The only problem is there is text stating "root".

It can be nice though to specify the user in the text as this does now.
Maybe we could look at the arguments or set an environment variable?